requirements.txt
Flask
Redis
flask-socketio
mysql-connector-python